CQB update
using getPos + doMove does not work for units inside houses, hence a path of all the buildingPositions is necesary.
Then use a looping function to force the unit to each position on the path.
However that will cause the unit to not engage visible targets.
In order to fix that a visible-enemy function was needed.

The init-CQB is spawned once map-objects are loaded

+Param for distance to start CQB
+Param for distance to activate CQB

14 new CQB functions (found in the CQB folder)
assignAllBuildings
assignBuilding
currentBuilding
unitIndoorPos
floorPositions
buildingFloors
unitsInBuilding
initCQB
CQBlimitReached
buildingPath
buildingExits
CQBTargetEnemies
clearCQBPos
clearBuilding